+++ otx/algorithms/classification/adapters/mmcls/task.py
@@ -29,6 +29,7 @@ from mmcls.utils import collect_env
 from mmcv.runner import wrap_fp16_model
 from mmcv.utils import Config, ConfigDict
 
+from otx.algorithms import TRANSFORMER_BACKBONES
 from otx.algorithms.classification.adapters.mmcls.utils.exporter import (
     ClassificationExporter,
 )
@@ -280,15 +281,16 @@ class MMClassificationTask(OTXClassificationTask):
             model.register_forward_pre_hook(pre_hook)
             model.register_forward_hook(hook)
 
+        model_type = cfg.model.backbone.type.split(".")[-1]  # mmcls.VisionTransformer => VisionTransformer
         if (
-            not dump_saliency_map or "Transformer" in cfg.model.backbone.type
-        ):  # TODO: remove "Transformer" in statement after resolving Issue#2098
+            not dump_saliency_map or model_type in TRANSFORMER_BACKBONES
+        ):  # TODO: remove latter "or" condition after resolving Issue#2098
             forward_explainer_hook: Union[nullcontext, BaseRecordingForwardHook] = nullcontext()
         else:
             forward_explainer_hook = ReciproCAMHook(feature_model)
         if (
-            not dump_features or "Transformer" in cfg.model.backbone.type
-        ):  # TODO: remove "Transformer" in statement after resolving Issue#2098
+            not dump_features or model_type in TRANSFORMER_BACKBONES
+        ):  # TODO: remove latter "or" condition after resolving Issue#2098
             feature_vector_hook: Union[nullcontext, BaseRecordingForwardHook] = nullcontext()
         else:
             feature_vector_hook = FeatureVectorHook(feature_model)
@@ -646,9 +648,7 @@ class MMClassificationTask(OTXClassificationTask):
     def _init_hparam(self) -> dict:
         params = self._hyperparams.learning_parameters
         warmup_iters = int(params.learning_rate_warmup_iters)
-        if (
-            self._multilabel and self._recipe_cfg.model.backbone.get("arch", True) != "deit-tiny"
-        ):  # TODO: resolve this hack
+        if self._multilabel:
             # hack to use 1cycle policy
             lr_config = ConfigDict(max_lr=params.learning_rate, warmup=None)
         else: