+++ otx/algorithms/classification/configs/deit_tiny/model_multilabel.py
@@ -2,7 +2,7 @@
 
 # pylint: disable=invalid-name
 
-_base_ = ["../../../../recipes/stages/classification/incremental.yaml", "../base/models/deit.py"]
+_base_ = ["../../../../recipes/stages/classification/multilabel/incremental.yaml", "../base/models/deit.py"]
 ckpt_url = "https://download.openmmlab.com/mmclassification/v0/deit/deit-tiny_pt-4xb256_in1k_20220218-13b382a0.pth"
 
 model = dict(