+++ otx/algorithms/classification/configs/deit_tiny/semisl/__init__.py
@@ -1,6 +1,6 @@
-"""Initialization of EfficinetNet-B0 model for Semi-SL Classification Task."""
+"""Initialization of deit-tiny model for Semi-SL Classification Task."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2023 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.