+++ otx/algorithms/classification/configs/deit_tiny/semisl/model_multilabel.py
@@ -1,4 +1,4 @@
-"""EfficientNet-B0 config for semi-supervised multi-label classification."""
+"""deit-tiny config for semi-supervised multi-label classification."""
 
 # pylint: disable=invalid-name
 
@@ -6,7 +6,7 @@ _base_ = ["../../../../../recipes/stages/classification/multilabel/semisl.yaml",
 ckpt_url = "https://download.openmmlab.com/mmclassification/v0/deit/deit-tiny_pt-4xb256_in1k_20220218-13b382a0.pth"
 
 model = dict(
-    type="SAMImageClassifier",
+    type="SemiSLMultilabelClassifier",
     task="classification",
     backbone=dict(arch="deit-tiny", init_cfg=dict(type="Pretrained", checkpoint=ckpt_url, prefix="backbone")),
     head=dict(